backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pydantic-settings==2.4.0
pydantic[email]==2.8.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9
cloudinary==1.41.0
slowapi==0.1.9
